refactor: remove most examples
The repo included a vast number of example scripts and tests
the package with both the tornado and standard lib circuits.
Remove all of them, since they are not needed for a regular
user of the program.
Further remove all the root level example files and python
scripts used to generate r1cs circuits from circom file.
We have to add the circom compiler as dependency in the
future pr s.t. we can generate example in a scalable way.